backend/requirements.txt
fastapi
uvicorn[standard]
httpx
PyMuPDF
langchain
sentence-transformers
faiss-cpu
numpy
python-dotenv
